Fix all clippy warnings across ffgl-core, ffgl-glium, and build-common
- Replace transmute with f32::to_bits/from_bits for float<->u32 conversions
- Convert Into impls to From impls (clippy::from_over_into)
- Add safety docs to unsafe functions
- Replace match-then-bool with matches! macro
- Use if-let instead of single-arm match (clippy::single_match)
- Remove redundant borrows (clippy::needless_borrow)
- Remove unused imports (thread, Duration in isf_validator)
- Remove redundant &'static on constants
- Fix .clone() on Copy type (FFGLViewportStruct)
- Use .as_deref() instead of .as_ref().map(String::as_str)
- Use .map_or() instead of .map().unwrap_or()
- Fix 0 as u32 → 0u32 unnecessary casts
- Change const Once to static Once (interior mutability)
- Mark handle_plugin_main as unsafe fn (raw pointer deref)
- Suppress all clippy lints on generated binding files
- Fix doc comment indentation

+++ ffgl-glium/src/lib.rs
@@ -209,8 +209,8 @@ unsafe fn blit_fb((read_w, read_h): (u32, u32), (write_w, write_h): (u32, u32)) 
     };
 
     let target_rect = BlitTarget {
-        left: 0 as u32,
-        bottom: 0 as u32,
+        left: 0,
+        bottom: 0,
         width: write_w as i32,
         height: write_h as i32,
     };
